PORT: Type cleanup in dtoa.c

+++ src/dtoa.c
@@ -647,7 +647,7 @@ multadd
 			Bfree(b);
 			b = b1;
 			}
-		b->x[wds++] = carry;
+		b->x[wds++] = (ULong)carry;
 		b->wds = wds;
 		}
 	return b;
@@ -841,7 +841,7 @@ mult
 				*xc++ = z & FFFFFFFF;
 				}
 				while(x < xae);
-			*xc = carry;
+			*xc = (ULong)carry;
 			}
 		}
 #else
@@ -1653,7 +1653,7 @@ rshift(Bigint *b, int k)
 			while(x < xe)
 				*x1++ = *x++;
 		}
-	if ((b->wds = x1 - b->x) == 0)
+	if ((b->wds = (int)(x1 - b->x)) == 0)
 		b->x[0] = 0;
 	}
 
@@ -1902,7 +1902,7 @@ gethex( CONST char **sp, U *rvp, int rounding, int sign)
 		word1(rvp) = Big1;
 		return;
 		}
-	n = s1 - s0 - 1;
+	n = (int)(s1 - s0) - 1;
 	for(k = 0; n > (1 << (kshift-2)) - 1; n >>= 1)
 		k++;
 	b = Balloc(k);
@@ -1931,7 +1931,7 @@ gethex( CONST char **sp, U *rvp, int rounding, int sign)
 		n += 4;
 		}
 	*x++ = L;
-	b->wds = n = x - b->x;
+	b->wds = n = (int)(x - b->x);
 	n = ULbits*n - hi0bits(L);
 	nbits = Nbits;
 	lostbits = 0;
@@ -2546,7 +2546,7 @@ strtod
 		else if (nd < 16)
 			z = 10*z + c - '0';
 	nd0 = nd;
-	bc.dp0 = bc.dp1 = s - s0;
+	bc.dp0 = bc.dp1 = (int)(s - s0);
 	for(s1 = s; s1 > s0 && *--s1 == '0'; )
 		++nz1;
 #ifdef USE_LOCALE
@@ -2570,13 +2570,13 @@ strtod
 #endif
 	if (c == '.') {
 		c = *++s;
-		bc.dp1 = s - s0;
+		bc.dp1 = (int)(s - s0);
 		bc.dplen = bc.dp1 - bc.dp0;
 		if (!nd) {
 			for(; c == '0'; c = *++s)
 				nz++;
 			if (c > '0' && c <= '9') {
-				bc.dp0 = s0 - s;
+				bc.dp0 = (int)(s0 - s);
 				bc.dp1 = bc.dp0 + bc.dplen;
 				s0 = s;
 				nf += nz;
@@ -3367,8 +3367,8 @@ strtod
 		else {
 #ifdef Avoid_Underflow
 			if (bc.scale && y <= 2*P*Exp_msk1) {
-				if (aadj <= 0x7fffffff) {
-					if ((z = aadj) <= 0)
+				if (aadj <= (double)0x7fffffff) {
+					if ((z = (ULong)aadj) <= 0)
 						z = 1;
 					aadj = z;
 					aadj1 = bc.dsign ? aadj : -aadj;
@@ -3944,7 +3944,7 @@ dtoa
 			 */
 			dval(&eps) = 0.5/tens[ilim-1] - dval(&eps);
 			for(i = 0;;) {
-				L = dval(&u);
+				L = (int)dval(&u);
 				dval(&u) -= L;
 				*s++ = '0' + (int)L;
 				if (dval(&u) < dval(&eps))
